backend/requirements.txt
django
gunicorn
django-heroku
dj_database_url

django-cors-headers

# Rest Framework dependencies
djangorestframework
markdown
django-filter
djangorestframework-jwt

# Social Authentication
rest-social-auth
# Rest social auth requires the following version
social-auth-app-django>=2.1.0,<3.0
